+++ packages/components/bolt-image/__tests__/image.js
@@ -33,4 +33,36 @@ describe('<bolt-image> Component', () => {
     expect(results.ok).toBe(true);
     expect(results.html).toMatchSnapshot();
   });
+
+  test('<bolt-image> with max-width of 200px renders as expected', async () => {
+    const results = await render('@bolt-components-image/image.twig', {
+      src: '/fixtures/1200x660.jpg',
+      alt: 'A Rock Climber',
+      max_width: '200px',
+    });
+    expect(results.ok).toBe(true);
+    expect(results.html).toMatchSnapshot();
+  });
+
+  test('<bolt-image> with max-width of 50% renders as expected', async () => {
+    const results = await render('@bolt-components-image/image.twig', {
+      src: '/fixtures/1200x660.jpg',
+      alt: 'A Rock Climber',
+      max_width: '50%',
+    });
+    expect(results.ok).toBe(true);
+    expect(results.html).toMatchSnapshot();
+  });
+  test('<bolt-image> with max-width and additional style attribute renders as expected', async () => {
+    const results = await render('@bolt-components-image/image.twig', {
+      src: '/fixtures/1200x660.jpg',
+      alt: 'A Rock Climber',
+      max_width: '200px',
+      attributes: {
+        style: 'display: block;',
+      },
+    });
+    expect(results.ok).toBe(true);
+    expect(results.html).toMatchSnapshot();
+  });
 });